requirements.txt
selenium
pytest
pytest-html
pandas
openpyxl
allure-pytest
